[Refactor]#107-BookSearchContainerView 생성
